requirements.txt
gunicorn
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.0
av==12.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
ctranslate2==4.4.0
cycler==0.12.1
et-xmlfile==1.1.0
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gradio==5.0.1
gradio_client==1.4.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
humanfriendly==10.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
numpy==2.1.2
onnxruntime==1.19.2
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.2
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
ruff==0.6.9
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
srt==3.5.3
starlette==0.38.6
sympy==1.13.3
tokenizers==0.20.1
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
websockets==12.0
spacy
deepmultilingualpunctuation
numpy==1.26.4

